Add a root parameter to cmdline when there's no initramfs
Normally we would put a configuration file into the initramfs, but
custom kernels may boot without initramfs, so it's useless for them. We
still need to specify a root for them, so put it in the cmdline.

Functions are from my flash-kernel merge request, with small changes.
Also, uses findmnt instead of reading /etc/fstab directly.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>