Make configuration simpler and more intuitive
We used to only check the first block for a machine, which meant small
user changes to a machine's configuration would mask other values in
the database. This commit implements a more intuitive interpretation of
these configuration files, where settings in later blocks override those
in earlier blocks.

Thanks to this we can also just concatenate the user config and machine
database in priority order and search through the merged file instead of
searching both individually.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>