Improve input and temp file handling in mkdepthcharge
This defines temporary files in a single place and allows environment
variables to override them. Since the temporary folder is deleted at
exit, we can use this to inspect outputs of individual steps.

Furthermore, this extends support for providing intermediate files (via
environment variables) so that we can skip some steps. It's not as good
as a Makefile though.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>