Change configuration and database format
Instead of reinventing configuration, use something similar to what
flash-kernel is doing for the machine database there.

Major differences are in how machine-detection is overridden and how the
kernel command line is set, we use a config file similarly formatted as
the db for both and more.

As something extra, we have config.d and userdb.d folders to avoid users
having to edit the conffiles and maybe getting a conflict on upgrade.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>