depthchargectl: Process partition sizes as well
This is to make the write command ignore partitions smaller than the given
image, as the image cannot be written to those. We warn the user if a
partition is ignored this way since they definitely have a partition
smaller than the maximum image size their device allows for, and they
might want to fix that.

Also makes the list command accept SIZE as a valid column printing the
partition size in bytes, but it is not printed by default.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>