depthchargectl: write: Don't overwrite current partition
Can be overrridden with --force, which will hopefully only used if the
user only has one ChromeOS kernel partition. Ideally, there should be
multiple such partitions which will be used in a rotation. After
flashing an image. a successful boot process would mark it successful;
but an unsuccessful boot would go back to the working image as it runs
out of tries and becomes zero priority. With a single partition, we'll
have overwritten the working image, so an external recovery method will
be necessary if the image is bad.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>