mkdepthcharge: Add support for zimage format
From what I can understand the depthcharge supports FIT images only on
ARM systems. On x86/amd64 it instead directly packs the 'vmlinuz' file.
See KERNEL_FIT and KERNEL_ZIMAGE kconfig options in depthcharge sources.

The vbuil_kernel program modifies the vmlinuz if it detects it's a Linux
kernel image, but only when invoked with --arch x86*. It doesn't do that
if called with --arch arm*.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>