db: Add commented-out guesses for many ARM machines
These machines are the results of running this on the Linux tree:

    grep -i -E '[^,]model = .*(google|chrome).*' arch/arm*/**/*.dts

For arch/arm/boot/dts/*.dts, guessed:

    Kernel-Compression: none
    Max-Size: 16777216
    Image-Format: fit

For arch/arm64/boot/dts/*/*.dts, guessed:

    Kernel-Compression: none lz4 lzma
    Max-Size: 335544332
    Image-Format: fit

"Google Scarlet" has two variants with the same machine name, so used a
wildcard in the DTB-Name to capture both. "Google Cheza" has multiple
revisions with different dtbs, but I added that with a wildcard as well.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>