depthchargectl: Add console=tty0 to default command line
The device-tree files are setting the default console to serial for ease
of debugging (access to early kernel messages):

    chosen {
            stdout-path = "serial2:115200n8";
    };

This makes initramfs messages and prompts (cryptsetup unlock, emergency
shell) to be sent to the serial console instead of the screen (unless
plymouth is also used in the initramfs). The easiest way to counteract
that issue is to set console manually in the command line.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>