mkdepthcharge: Print info from vbutil_kernel in verbose mode
I like that mkimage prints information about the built image whenever it
does so, but vbutil_kernel is silent. This runs vbutil_kernel --verify
after building the image to provide info, but only in verbose mode. At
the very least it shows the full kernel command line which is useful to
know.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>