depthcharge-tools: Add python packaging stub
I started writing all this with the assumption that I'd have to run it
in debian-installer and thought I had to choose between C, dash shell
and perl. But now I know I don't have to, so I'm rewriting it in python.

It may be necessary to keep updating the shell version, so I'm reserving
v0.4 for something like that. Calling the python version v1.0 is a big
leap, so let's call it v0.5 for now.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>