mkdepthcharge: Set argument defaults
Set the --arch to host machine's by default. For --name, mkimage uses
"unavailable" for the other images, so default to that. For --compress,
we need to pass "none" or mkimage defaults to "gzip".

I can't set --cmdline default in the parser since it would append to the
default, and other argument defaults depend on the values for yet other
arguments so check all those after parsing.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>