mkdepthcharge: Fix entry point argv logic
I thought the console scripts entry point would be called by passing
sys.argv[1:] into it, but that's not the case. Here, main() is being
called with no arguments and should itself read from sys.argv.

The parse_args call was also wrong, but happened to work since the part
in __main__ was equally wrong in the opposite way and balanced it out,
so fix that too.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>