depthchargectl: Fix --verbose as a global argument
Since I have the same "store_true" argument in the parser and the
subparser, the default (False) in the subparser is set after and
overrides the given argument in the parent parser.

Fix that by setting the correct value as the subparsers' default when
the parent parser encounters the option.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>