depthchargectl: Fix printing all partitions for nonexistent args
When given a nonexistent file as an argument, find_disks() would return
an empty list. Then, when we call "cgpt find" with that empty list we get
all the partitions cgpt find could find, which is not what we wanted.
Fix it by replacing the empty list with the original list.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>