utils.py: Refactor disk and partition handling
The main problem is disk images files and partitions contained in them.
As those partitions have no block device corresponding to them outside
the image, Partition has to be a non-Path class that takes a disk and a
partition number. But it should still work when given a partition block
device. This is a first attempt at getting things to work.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>