utils.py: Split disk and partition handling by a BlockDevice
This is an attempt to unify handling of the block-device partitions and
disk-image partitions. The partitions command would handle its arguments
as block devices or disk images, where block devices are created as
either partition or disk block devices from a single constructor (via
BlockDevice.__new__). Then, regardless of that distinction we can get
their parent disks, combine those with disk images, and pass the entire
thing to "cgpt find" which would also not care for their type and get
partitions by the same method.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>