utils.py: Don't subclass Path for disk and partition classes
Subclassing Path has a fatal flaw in that Path operations that would
result in Path objects now result in objects of the subclass, i.e.
Disk("/dev/mmcblk0").parent is Disk("/dev") which is obviously wrong.

Instead of trying to cover everything pathlib.Path could do, keep the
path as a property of the relevant classes.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>