utils.py: Don't split classes on them being a block device or not
Now that I don't subclass Path, I don't need different classes for
block-device partitions and disk-image partitions (which are just
partitions with and without a concrete path). I can just use the
Partition.path property for the difference and set it to None if a file
for the partition doesn't exist.

In hindsight, subclassing Path has been a gigantic waste of time.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>