depthchargectl: Implement printing the partition table
With this the partition command is mostly done. I can't use getattr to
access the Partition properties because I can't get p.disk.path easily
with that. At the end, I can't simply join the columns with tab
characters because it doesn't ensure alignment, so I calculate the
minimum necessary width of the columns and print each value formatted to
that width of its column. To keep things good looking, hard-code 4 as
an absolute minimum column width.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>