depthchargectl: Implement most of the target subcommand
With the rest of the helpers in place, the target command is relatively
straightforward. We sort and filter the given arguments' partitions, then
print the least-important one.

The argument parser was only taking at most one argument as the target
partition, fix that by taking all arguments as disks and only showing
the partition support in the help message.

Also add helpers to get the currently booted partition for checking the
target partition against it.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>