utils: Add basic machine db and config parsing
The format of these files were based on flash-kernel's machine db file.
It's quite easy to parse them with configparser if we add some hacks to
make it consider "Machine: <machine>" lines in the db as sections and
consider the config file as if it's part of the DEFAULT section.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>