utils: Add kernel_cmdline() to read /proc/cmdline
Split it with shlex so that arguments like a="b c" are understood
correctly. However, that ends up as ['a=b c'] which is a bit incorrect
as the quotes are missing, but useful as I can use str.partition("=") to
get the value without quotes. Not sure if it'd be good to make a Cmdline
class.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>