utils: platform: Add comparison ops to Kernel by release versions
This compares the versions by splitting the parts by dashes and dots,
then building a special list from the parts that would be compared
instead. It's very ugly because the need to sort "rc" and Debian's
"trunk" prerelease versions before the proper version, and to sort
numbers as numbers.

Sorts numbers before non-numbers for now, not sure if that's
appropriate.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>