utils: configparser: Convert config and board info parsing to classes
Primarily so that I can define and use properties for each config and
board info option. I considered making them subclasses of ConfigParser,
but I would need to do too much wrapping to make BoardInfo a container
of Boards and to make Config act as a section while being a config.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>