depthchargectl: rm: Add initial implementation
Can remove flags from partitions for the given image and remove the
images from the var/ directory, which is enough for now. Reading the
first 0x10000 bytes of the partition should be enough as it contains the
entirety of the vblock header including a signature on the image
contents. (Running vbutil_kernel with --vblockonly creates a file of
size 64KiB == 0x10000).

Also combine both "kernel-version" function argument into "image", while
keeping it in the help message.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>