mkdepthcharge: Fix checks for compress and name variables
These variables were being set to defaults for FIT, but the checks for
zimage expected them to not be set. Make setting those defaults
conditional on image type so the check works correctly.

Also, when called with "--format zimage", complain loudly at "--compress
none" and "--name unavailable" instead of accepting and doing nothing
with them.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>