depthchargectl: target: Add comments and logging messages
This copies what we can from the shell version, but things are quite
different than the shell version in their semantics, mainly in checking
validity of a partition and exiting with a status code for whatever
condition it fails. Ideally, that would be done in the Partition helper
class and an error would be raised there for partition validity. Though
things like checking partition size and currently booted partition is
appropriate here.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>