depthchargectl: rm: Don't check argument in installed kernels
The rm subcommand might be run (automatically) after a kernel is
uninstalled, the uninstalled kernel wouldn't show up and its partitions
wouldn't be disabled and images wouldn't be removed. The shell version
already did this the correct way, but I had just forgot to consider it
in my hasty reimplementation.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>