depthcharge-tools: Configure DATADIR, SYSCONFDIR etc. at build time
This reads things from environment (or falls back to some defaults) and
puts the important dirs into the built package as a resource so we can
read them in __init__.py.

Right now this can not distinguish between develop/normal or between
system/user installations, but it'll need to do that to use the correct
files and paths for each install type.

I think I'll eventually merge this dirs file into the config file and
include that as a resource as well, but I'm postponing that a bit since
I want things to be able to work the old way for now.

Also, I guess this could be injected into the setuptools configure step,
but didn't figure all that out yet.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>