setup.py: Install files relative to PREFIX
When these data_files directories are absolute paths, they are still
interpreted as relative to package installation directory when running
`PREFIX=~/.local pip3 install --user`, so data files get installed to:

    ~/.local/lib/python3.9/site-packages/home/alpernebbi/.local/share/...

Passing them as relative to the PREFIX gets them into the correct paths
in that case. However it probably means that system-wide installations
will always install as if PREFIX=sys.prefix, as the relative paths will
be interpreted as relative to that.

Also break the file entries into lines for a more consistent look.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>