depthcharge-tools: Fix git-based version number
Directly using the git-describe output results in a LegacyVersion as
the only thing that can come after .dev<N> is a +<local> version.
Convert the git describe parts into dot separated parts as this local
version (like 0.5.0.dev1+5.g93f716b52b94) to fit the spec in PEP440.

Also, try to derive version from git even when the package is installed,
so that we can have the git-version in editable installs.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>