depthcharge-tools: Use config.ini resource as a string
I don't really need a filename for this file, and it looks like that
could fail in some conditions like running the package from a zip file.
However, pkg_resources.resource_string returns bytes and needs decoding.

I tried using resource_stream, but then it opens the file as text and
ConfigParser._read() complains that it's not bytes...

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>