utils: Use board codenames to identify boards
This makes platform.board_name() return the board's codename instead of
the devicetree model, and makes Config get the board's config based on
the codename instead of config section headers.

Using the devicetree model for this was originated from flash-kernel.
But it's not really appropriate since this is an arbitrary string maybe
subject to change, and there are boards that won't have a devicetree.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>