config: Use hierarchical section names for boards
I'm planning on doing some inheritance with boards, so I need a way to
express the "parent" of a board. The section names seems the best for
this, e.g. boards/arm64/gru/kevin -> boards/arm64/gru -> boards/arm64.

The section names were used as printable names for the boards, so move
that to a name key in the section. Instead of devicetree model string,
use the name field in Chrome OS's recovery.conf for existing devices.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>