config: Implement inheritance for config sections
This implements a custom dictionary that finds all parents for a key,
combines their values in topological order to a ChainMap and returns
that when the values are dictionaries.

We can't do this in our wrapper Config class since ConfigParser prefers
default values to fallback values, making it impossible to determine
whether a value retrieved from a child section is set in the child
section or the default section.

Now that inheritance works, remove values from the config file which can
be inherited from parent boards.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>