utils: argparse: Improve Argument nargs deduction
Arguments with defaults should be considered optional. However argparse
has no way to to express that we want a number of arguments within a
range. For those let's ask for all we can get, and check the result
later.

This doesn't accept keyword arguments for now, but it should be doable
in a format like "--arg key1=val1 key2=val2".

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>