utils: argparse: Fix partial's doc leaking to the help message
Since we moved to the custom action, we have been using the partial
function's docstring which isn't the same as the wrapped method's
docstring, as we weren't actually wrapping it. Once we wrap it, we need
to tell inspect.signature to not descend to the wrapped function to
avoid including the bound self in the lists of parameters.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>