utils: argparse: Add initial implementation of groups
This works good enough for now, but copies a lot from Argument which
I might try to refactor. There is also the issue of decorating a method
with multiple Argument()s, I thought it would be done with creating an
ad-hoc Group, but that causes problems with the help message.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>