utils: argparse: Implement Command subcommands
Not sure about this implementation but it works for now. Command objects
set as another Command's class attributes become its subcommands.
However this has no way to customize the add_subparsers() arguments.

Temporarily disable the owner check for multiple owners, right now the
subparser Argument and Groups appear to be owned by the Command class
instead of an instance, and trying to reset the owner to the subcommand
instance would raise an error.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>