utils: argparse: Make ArgumentParser kwargs accessible from Commands
It's great to have these to know what will be passed to argparse.
However they shadow the Command attributes with the same names because
they're descriptors on the metaclass. We can't modify type __dict__s,
so we need to temporarily unset the property and re-set it to pretend
it's not there when the attributes are used as such.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>