utils: argparse: Use inspect.BoundArguments for parsing results
This is a great unique type that has bound .args and .keywords easily
accessible, and I can check the type to see if I should call the
argument methods or not.

Only use it for arguments that have wrapped a method and let others
either fall back to "store" or specify their own action as I don't want
to fake a function signature by nargs etc.

With that it makes more sense to call it FunctionBindAction instead of
ArgumentAction. And I have been irritated by it's position between
Action and Group, Command etc. so move it above.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>