utils: argparse: Create class-bound copies for method decorators
Without this the __self__ attribute just referred to the first getattr
that was done, which is a bit weird (cmd.sub.arg.__self__ could be
either cmd or cmd.sub). Create a bound object each time we access the
attribute so it's always the correct object (cmd.sub for the above).

Skipping group arguments while building the parser implicitly depended
on there being a single Argument object for the Command, so this
would be duplicating them. Also fix that by making copied Arguments
remember which group they belong to.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>