utils: argparse: Use Command instance dict for Argument method bound args
Since our custom argparse Action only returns inspect.BoundArguments(),
we can check and call the bound methods only when the argument is set to
that type. That means instead of keeping track of what inputs we were
given and what values we should return from __get__ in the Argument
instance, we can offload that to the Command instance __dict_.

As we don't need a Command instance bound Argument object, we can
instead replace returning that with returning the Argument's __func__
bound to the Command instance to make the instance look slightly more
like an ordinary class.

Also remove the Argument._unset whose sole purpose was to help implement
the Argument's .inputs and .value that this replaces.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>