utils: argparse: Make Groups call their methods on BoundArguments
This is mostly for pseudo-Groups that get created by chaining Argument
decorators, for which the parser results will include the attribute name
of those Groups which will be set a BoundArguments object.

While we're at it, clean the Argument.wrap() a bit. Make those pseudo
Groups wrap the __func__ of their arguments to make the implementation a
bit easier, and stop doing manually the things Group.wrap() would do.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>