utils: argparse: Improve Command __call__ value binding
For methods that take a single value it should be possible to pass the
single value instead of a list containing that single item. For methods
that take no arguments, use truthiness of the argument to either call or
not call the method. Of course, if we get BoundArguments just pass that
along assuming the caller knows what they're doing.

Still not sure how to do keyword arguments, so skip that implementation;
but calling with those is still possible via BoundArguments.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>