utils: argparse: Add Command.main to parse args and run command
Having something like this is useful as a console_scripts entry_point.
Mostly the same as the one in OldCommand, but we need to remove None
arguments because it usually represents the lack of the command line
argument (Action.default=None) but __call__'s way of representing that
is simply not passing the keyword argument.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>