utils: argparse: Add append, count to FunctionBindAction
These are intended to replace the use cases for action="append",
action="count" behaviour of argparse.

Unlike argparse which calls the Action each time the argument is
encountered, our append=True collects the arguments from multiple
invocations to pass them to a single method call. As argparse sees the
option again and again it tries to consume a nargs number of command
line arguments, so only nargs="*" really makes sense with ours.
E.g. "-f a b -f c" would not work with f(a, b, *c).

The count=True is a bit weird because it needs nargs=0 but also a
function that can accept the resulting value in one of it's arguments,
so we need to adjust the Argument.nargs in a weird way.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>