utils: argparse: Let Arguments pre-bind arguments to methods
Callers can bind argument by either passing the "args", "kwargs" as
keyword arguments (most useful if the names conflict with the ones for
Action) or by passing the keyword arguments themselves directly into the
Argument constructor. This is mostly to enable something like:

    @Argument("--log-level")
    @Argument("--debug", level=5)
    @Argument("-q", "--quiet", level=-1)
    @Argument("-v", "--verbose", count=True)
    def loglevel(self, level):
        level = 50 - (int(level) * 10)
        logger.setLevel(level)
        return level

However pre-binding for count and append is not done as I couldn't think
of a good way of doing it in the occurrence of different Arguments asking
to pre-bind different sets of arguments multiple times.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>