utils: argparse: Try calling every Argument, Group with default args
Until now I had thought the methods shouldn't be called unless the
relevant argument was given on the command line, but that doesn't make
sense if I want the attributes to have the correct default values in
absence of the argument. The argparse "default" argument could be used
for that, but then I can't make it dependent on the values of other
arguments.

If the Argument / Group method is callable without any args (including
those with all having default values), call it even if the argument
isn't on the command line so that it can dynamically set a default
value.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>