depthchargectl: Rename modules, commands to avoid shadowing list
Having it as depthcharge_tools.depthchargectl.list somehow makes "list"
refer to the module object for that when I use it in commands. Rename
the subcommand modules to have an underscore in the beginning to avoid
that.

Since the names of the classes don't really matter, change them to the
depthchargectl_{cmd} format, and stop importing them by name.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>