utils: argparse: Catch AttributeError from wrapped methods
Raising AttributeError in __get__ makes getattr(attr, default) calls
return the "default" value as it's interpreted to mean the attribute
doesn't exist, but we want the AttributeErrors from inside the wrapped
methods to stop execution as that would be something wrong with the
code instead. Fix by catching AttributeErrors and raising RuntimeError
from those so that doesn't get caught.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>