depthchargectl: build: Drop kernel-compression config option
This was only used to avoid the warning when the resulting image size is
too big by e.g. excluding "none" from the compressions. It's not
something that should need to be configured, which compression we want
can most likely be deduced from the sizes of the input files, or we can
just build multiple times and check as we are currently doing.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>