depthchargectl: check: Don't check signature without public key
Due to buggy inheritence in utils.argparse helpers the Group method that
would set the vboot keys by default to OS keys does not run in the
subcommands. Disabling the signature checks here lets me postpone fixing
that.

(This is not a problem in the "build" subcommand only because
mkdepthcharge defaults to OS keys anyway.)

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>