mkdepthcharge: Add --tmpdir argument to keep temporary files
I think this would be useful for debugging, but the main idea here is
to let callers (e.g. depthchargectl) who have their own temporary dirs
pass in a subdirectory of that.

It also lets me refactor more of the input file handling into the
relevant arguments, but I'm not sure if I should be doing that.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>