depthchargectl: build: Choose dtbs based on dt-compatible
Running fdtget over all the dtbs (273 files on my system) takes a few
seconds but it's bound to make things more robust compared to using
globs for the filename. It also lets me stop caring about dtb-name, so
remove it from the configs.

Setting the config to "True" includes all dtbs, and "False" indicates dtb
files are not necessary for the board.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>