utils: os: Make DiskGraph use and return Disk, Partition objects
It uses Disk and Partition in its __init__, so we need to initialize it
after the definitions for those. Assign it to a "system_disks" variable
as "disks" as a variable is used a lot in the depthchargectl subcommands.
Rename the class to Disks as that makes more sense now.

Also make Disks._edges a dict instead of a defaultdict(set) to prevent
unintended additions as a result of accessing the dict.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>