utils: platform: Simplify kernel release version comparisons
This still tries to match Debian kernel version naming, which might not
be applicable to all distributions, but it should mostly work as
expected if they only use dots and dashes.

One caveat is that it would interpret git hashes as multiple parts, but
if the version number is same to the point where we're comparing hashes
the comparison is already undefined behaviour.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>