utils: argparse: Add CommandExit to return and exit with status code
Until now, we were piggybacking on OSError to set exit status for
Command.main(). But OSError and its subclasses and their errno values do
not really correspond to what we want the commands to exit with. This
tries to implement a similar exception-based mechanism to specify exit
statuses for commands.

One point of interest is commands which would want to return a value
(like True or False) when called within Python, but set an exit status
(0 or 1) when called as main. Consequently, they cannot rely on raising
an Exception to communicate the exit status.

Deciding the exit status by processing the returned value per se doesn't
make sense as it's not possible to predict the entire output space.
Setting a custom attribute on the Command instance wouldn't work as the
instance is temporary in the CommandMeta.__call__() scope and
inaccessible to Command.main().

So the only option I see viable with that constraint is returning a
custom object that represents both the output value and the exit status,
which I've chosen to be the same exception that can be raised as a
replacement for OSError.

Still, CommandMeta.__call__() needs to know whether it should unwrap the
CommandExit to return the intended Python value, or pass it unwrapped so
Command.main() can know about the exit status. This is done by passing a
private keyword argument which indicates it should not unwrap the value.

To unify code with the other case where CommandExit would be raised,
CommandMeta.__call__() also raises this Exception instead of returning
it.

All this is ugly and hacky, but I can't think of a better way to do it
for now.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>