depthchargectl: check: Raise custom CommandExits instead of OSError
Move the image.is_file() check to the Argument because it fits there
better. This makes it a parser error with status 1 instead of returning
status 2 as it had been doing so far, but I'm not sure if I still want
to return different status codes for different cases anymore.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>