utils: os: Fix Partition path being set to disk path
When initialized as Partition(diskpath, partno) for disk images, the
newly created object would end up having the path attribute set to the
value given in the constructor meant to be the path for the disk. Fix by
setting the partition path to None after interpreting it as the disk
path.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>