utils: argparse: Print traceback for Command errors on DEBUG loglevel
I have been using the "raise CommandExit from Exception" form for a few
things, and it's useful to be able to print those exception tracebacks,
but only on the most verbose loglevels. Do not print the traceback when
returncode is 0 as that is usually not an error but a normal return.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>