depthchargectl: write: Pass configured options to other command calls
This does not pass verbosity as it would keep increasing log granularity
for the entire run call with every subcommand call.

Although the subcommands can read the values for their own sections,
their global config values are overwritten with the one for the write
call due to being set in the subcommand invocation.

It looks like a dict containing these could be made available as e.g.
depthchargectl.config_options attribute, but I'm postponing diving into
utils.argparse code as far as possible.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>