depthchargectl: write: Refactor positional argument handling
The input argument is primarily interpreted as a kernel version, and
used as a path to a file only if that interpretation fails. The kernel
version might not have an existing image, so we can't check this via
paths in images-dir, so we have to do the search here instead of
delegating it to the build subcommand.

Also avoid printing the KernelEntry object by removing the kernel
version log message as that will be printed by the build subcommand
anyway.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>