depthcharge_tools: Move stderr logging initialization to Command.main
The logging configuration should entirely be the responsibility and
at the discretion of the top-level application, importing a python
library should not cause log configuration to change in any way.
Avoid changing it by initializing the stderr log handler only in the
Command.main() call, where we know this invocation is the top-level
application.

Also add a top-level NullHandler to prevent the handler of last resort
from printing warning logs to stderr in absence of any configuration.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>