depthcharge_tools: Make verbosity options not modify the root loglevel
Now that we have logger instances for each Command (and solved some
other logging infrastructure problems) we can make the verbosity options
set loglevels on the Command-specific loggers. We can also pass it to
other command calls.

Also set default loglevels to WARNING as Python does by default on the
root logger, and make the Argument return the count of --verbose flags
instead of the loglevel to make it easier to pass to other commands.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>