depthchargectl: list: Return a custom UserList object
I don't think it will be necessary to ever run this command in Python,
as it's use cases can be covered by utils.os objects. However it
returning the table as a string is wrong in principle, so make it return
a list-like object whose string representation is that table.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>