depthcharge-tools: Add update_config.py command skeleton
I think I should automate the config.ini board sections as much as
possible, and all the fixups/quirks for all the things I'll be analyzing
should go into a file. So far I've been using interactive sessions and
an ad-hoc script file with hardcoded paths to experiment, but I believe
it all can be more clean and organized if I turned it into a Command
instead.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>