update_config.py: Make inputs optional
Board-overlays, depthcharge, coreboot are available per release branch
but the recovery.conf file and the chromiumos/project repo are not. I
think it would be better to iterate through the release branches for the
former and apply the latter at the end, so it makes sense for these to
be optional.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>