depthchargectl: Fix using sub-board sections for a parent codename
I had expected the codename-in-section check to filter out sub-boards,
but it returns True for them as well. Even if I fixed it in the
ConfigDict, I'll also have to check multiple hwid-match values for a
single codename which means mapping codenames to a single board won't
really work here. Luckily I wasn't depending on codename that much, so
I can change those to section names.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>