depthchargectl: Fix passing SectionProxy as config argument
Directly using the SectionProxy here makes us ignore the values set in
the callee's config section even if they're unset in the passed section.
Instead, update the callee's section based on the latter.

For example, the write subcommand was passing an empty section to the
build subcommand, causing it to not have a kernel-cmdline value although
a default is set for build. This commit makes that build invocation
correctly use that default value.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>