depthchargectl: bless: Run services only on cros_secure boots
When booted with e.g. u-boot in RW_LEGACY, we most likely won't have a
kern_guid in the command line (we won't have %U expansion even if we
did) and we definitely won't care which ChromeOS Kernel partiiton should
be blessed as we wouldn't be using them. So don't even try blessing
then, based on the "cros_secure" kernel cmdline arg depthcharge injects.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>