depthcharge_tools: Adjust some log messages
Confirmations that a command did something should be printed to the user
in the default verbosity, which means I have to raise them to the
warning level although they're not really warnings.

Also the info log messages for the mkdepthcharge input files were
getting printed on the default loglevel because they were being called
before the verbosity argument, put them into the __call__ so that they
don't get inappropriately printed.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>