utils: subprocess: Add methods for fdtget
I think this is mostly going to be useful for the check subcommand. It's
a bit slow while ripping out e.g. a kernel image from a FIT image, but
running fdtget itself is slow on my system for that case. It may be
better to integrate with pylibfdt or something, but for now do whatever
works.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>